first attempt at separating out windows/unix functionality
